[UPDATE] update size_t to int in printing function

+++ sorting/C/utils.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "utils.h"
 
-void print_array(int *arr, size_t length) {
+void print_array(int *arr, int length) {
     printf("arr[");
     for (int i = 0; i < length - 1; i++) {
         printf("%d, ", arr[i]);
